requirements.txt
fastapi
uvicorn
requests
python-multipart
pymupdf
python-dotenv